Fix how the clients (both ec2 and route53) are being created.
Previously, the ec2 client was being created during require('./lib/aws.js')
but this was _before_ the AWS_* env vars were being checked. This is fixed.

Also, we were also checking the valid regions async via the API but AwsSum
already knows what regions are allowed, so this is now sync.